Removing the new file.
